REF: Move related scripts into one directory
