+++ scripts/check_inter_service_compatibility_across_versions.py/process_question.py
@@ -4,9 +4,9 @@ import os
 import sys
 import tempfile
 
-from octue.resources import Datafile, Dataset, Manifest
+from utils import ServicePatcher
 
-from ..utils import ServicePatcher
+from octue.resources import Datafile, Dataset, Manifest
 
 
 try:
