+++ scripts/check_inter_service_compatibility_across_versions.py/utils.py
